fixed update and input
